requirements.txt
# API and data handling
python-binance==1.0.16
pandas==2.0.3
numpy==1.24.3
ccxt==3.0.74

# Technical analysis
pandas-ta==0.3.14b0

# Data visualization (for debugging and analysis)
matplotlib==3.7.2
seaborn==0.12.2

# Machine learning components (for AI signal filtering)
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1

# Utilities
python-dotenv==1.0.0
schedule==1.2.0

# Web interface
flask==2.3.3
